Move Route.jsx into the component folder
