+++ client/js/index.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from './store';
-import Routes from './Routes.jsx';
+import Routes from './component/Routes';
 import styles from '../sass/styles.scss';
 
 const app = document.getElementById('app');
